scripts/amazon/requirements.txt
python-amazon-paapi>=6.0.0
gspread>=6.0.0
google-auth>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
Pillow>=10.0.0
